backend/ai_parser.py
"""
AI Parsing Engine — Groq-Powered Skill Extractor (Llama 3)
Extracts skills from Resume and Job Description text,
then maps them STRICTLY to nodes in our SkillGraphEngine.
Zero-hallucination guarantee: output is validated against the graph.
"""

import os
import json
from groq import Groq
from dotenv import load_dotenv
from skill_graph import SkillGraphEngine

load_dotenv()


class AIParsingEngine:
    def __init__(self, skill_graph: SkillGraphEngine):
        self.client = Groq(api_key=os.getenv("GROQ_API_KEY"))
        self.model = "llama-3.3-70b-versatile"
        self.skill_graph = skill_graph
        self.valid_skill_ids = {s["id"] for s in skill_graph.get_all_skills()}
        self.skill_lookup = {
            s["id"]: s["name"] for s in skill_graph.get_all_skills()
        }

    def _build_skill_catalog_prompt(self) -> str:
        """Builds the strict skill catalog the model must choose from."""
        all_skills = self.skill_graph.get_all_skills()
        lines = []
        for s in all_skills:
            lines.append(
                f'  - id: "{s["id"]}" | name: "{s["name"]}" '
                f'| domain: {s["domain"]} | level: {s["level"]}'
            )
        return "\n".join(lines)

    def _call_llm(self, prompt: str, max_tokens: int = 1000) -> str:
        """Central LLM call — easy to swap model here if needed."""
        response = self.client.chat.completions.create(
            model=self.model,
            max_tokens=max_tokens,
            temperature=0.1,
            messages=[
                {
                    "role": "system",
                    "content": "You are a precise skill extraction engine. You always respond with valid JSON only. No markdown, no explanation, no backticks."
                },
                {
                    "role": "user",
                    "content": prompt
                }
            ]
        )
        return response.choices[0].message.content.strip()

    def _safe_parse_json(self, raw: str) -> dict | None:
        """Safely parses JSON, stripping markdown fences if present."""
        # Strip markdown fences
        if "```" in raw:
            parts = raw.split("```")
            for part in parts:
                part = part.strip()
                if part.startswith("json"):
                    part = part[4:].strip()
                if part.startswith("{"):
                    raw = part
                    break

        # Find first { to last } in case of extra text
        start = raw.find("{")
        end = raw.rfind("}") + 1
        if start != -1 and end > start:
            raw = raw[start:end]

        try:
            return json.loads(raw)
        except json.JSONDecodeError:
            return None

    def extract_skills_from_resume(self, resume_text: str) -> dict:
        """
        Parses resume text and returns matched skill IDs from our graph.
        Uses strict grounding — model can ONLY pick from the catalog.
        """
        catalog = self._build_skill_catalog_prompt()

        prompt = f"""Analyze the resume below. Identify which skills from the APPROVED SKILL CATALOG the candidate already possesses.

APPROVED SKILL CATALOG (select ONLY from these exact IDs):
{catalog}

RULES:
1. Only select skill IDs that genuinely match the resume content.
2. If candidate shows intermediate Python, select BOTH python_basics AND python_intermediate.
3. Be conservative — only include skills with clear evidence.
4. Include brief evidence for each skill (max 10 words).

Respond with this exact JSON structure:
{{
  "matched_skills": [
    {{"skill_id": "python_basics", "evidence": "3 years Python development experience"}},
    {{"skill_id": "git_basics", "evidence": "Used GitHub for version control"}}
  ],
  "candidate_summary": "One sentence summary of candidate profile",
  "experience_level": "junior",
  "primary_domain": "programming"
}}

RESUME:
{resume_text[:3000]}"""

        raw = self._call_llm(prompt)
        parsed = self._safe_parse_json(raw)

        if not parsed:
            return {
                "matched_skills": [],
                "known_skill_ids": [],
                "candidate_summary": "Could not parse resume.",
                "experience_level": "unknown",
                "primary_domain": "unknown",
                "hallucination_guard": {"removed_count": 0, "removed_ids": []}
            }

        # ── GROUNDING VALIDATION ──────────────────────────────────────
        validated = [
            item for item in parsed.get("matched_skills", [])
            if item.get("skill_id") in self.valid_skill_ids
        ]
        hallucinated = [
            item.get("skill_id") for item in parsed.get("matched_skills", [])
            if item.get("skill_id") not in self.valid_skill_ids
        ]

        return {
            "matched_skills": validated,
            "known_skill_ids": [s["skill_id"] for s in validated],
            "candidate_summary": parsed.get("candidate_summary", ""),
            "experience_level": parsed.get("experience_level", "unknown"),
            "primary_domain": parsed.get("primary_domain", "unknown"),
            "hallucination_guard": {
                "removed_count": len(hallucinated),
                "removed_ids": hallucinated
            }
        }

    def extract_skills_from_jd(self, jd_text: str) -> dict:
        """
        Parses Job Description and returns target skill IDs
        the role requires, with priority weighting.
        """
        catalog = self._build_skill_catalog_prompt()

        prompt = f"""Analyze the job description below. Identify which skills from the APPROVED SKILL CATALOG are required for this role.

APPROVED SKILL CATALOG (select ONLY from these exact IDs):
{catalog}

RULES:
1. Only select skill IDs genuinely required by this job description.
2. Assign priority: "must_have" or "nice_to_have" for each skill.
3. Include brief justification for each skill (max 10 words).

Respond with this exact JSON structure:
{{
  "required_skills": [
    {{"skill_id": "python_intermediate", "priority": "must_have", "reason": "Core language for backend development"}},
    {{"skill_id": "docker_basics", "priority": "nice_to_have", "reason": "Mentioned in preferred qualifications"}}
  ],
  "role_summary": "One sentence summary of the role",
  "seniority_level": "mid",
  "primary_domain": "ai_ml"
}}

JOB DESCRIPTION:
{jd_text[:3000]}"""

        raw = self._call_llm(prompt)
        parsed = self._safe_parse_json(raw)

        if not parsed:
            return {
                "required_skills": [],
                "target_skill_ids": [],
                "must_have_ids": [],
                "nice_to_have_ids": [],
                "role_summary": "Could not parse job description.",
                "seniority_level": "unknown",
                "primary_domain": "unknown",
                "hallucination_guard": {"removed_count": 0, "removed_ids": []}
            }

        # ── GROUNDING VALIDATION ──────────────────────────────────────
        validated = [
            item for item in parsed.get("required_skills", [])
            if item.get("skill_id") in self.valid_skill_ids
        ]
        hallucinated = [
            item.get("skill_id") for item in parsed.get("required_skills", [])
            if item.get("skill_id") not in self.valid_skill_ids
        ]

        return {
            "required_skills": validated,
            "target_skill_ids": [s["skill_id"] for s in validated],
            "must_have_ids": [
                s["skill_id"] for s in validated
                if s.get("priority") == "must_have"
            ],
            "nice_to_have_ids": [
                s["skill_id"] for s in validated
                if s.get("priority") == "nice_to_have"
            ],
            "role_summary": parsed.get("role_summary", ""),
            "seniority_level": parsed.get("seniority_level", "unknown"),
            "primary_domain": parsed.get("primary_domain", "unknown"),
            "hallucination_guard": {
                "removed_count": len(hallucinated),
                "removed_ids": hallucinated
            }
        }

    def generate_reasoning_summary(
        self,
        resume_result: dict,
        jd_result: dict,
        pathway_result: dict
    ) -> str:
        """
        Generates a human-readable explanation for WHY
        this specific pathway was chosen.
        Powers the Reasoning Trace feature (10% of judging score).
        """
        known_skills = [s["skill_id"] for s in resume_result.get("matched_skills", [])]
        target_skills = jd_result.get("target_skill_ids", [])
        gap_count = pathway_result.get("gap_count", 0)
        total_hours = pathway_result.get("total_hours", 0)
        trace = pathway_result.get("reasoning_trace", [])

        prompt = f"""You are an expert learning & development coach explaining a personalized training plan to a new hire.

CANDIDATE PROFILE:
- Summary: {resume_result.get('candidate_summary', 'N/A')}
- Experience Level: {resume_result.get('experience_level', 'N/A')}
- Known Skills: {len(known_skills)} matched
- Primary Domain: {resume_result.get('primary_domain', 'N/A')}

ROLE REQUIREMENTS:
- Role: {jd_result.get('role_summary', 'N/A')}
- Seniority: {jd_result.get('seniority_level', 'N/A')}
- Required Skills: {len(target_skills)}

PATHWAY STATS:
- Skill Gaps: {gap_count}
- Total Training Hours: {total_hours}
- Algorithm Trace: {json.dumps(trace)}

Write exactly 3 paragraphs (no bullet points, no headers):
1. The candidate's existing strengths and what the analysis found
2. Why these specific modules were chosen, referencing the algorithm trace naturally
3. What the candidate will be capable of after completing this pathway

Be specific, encouraging, and professional."""

        raw = self._call_llm(prompt, max_tokens=600)
        return raw