backend/requirements.txt
### File 2: Create `backend/requirements.txt`
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
groq==0.9.0
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
networkx==3.4.2
numpy==2.2.6
python-dotenv==1.0.0